# crates/kino-server/openapi.json
{"openapi":"3.1.0","info":{"title":"Kino API","description":"","license":{"name":"Apache-2.0","identifier":"Apache-2.0"},"version":"0.1.0-phase-2"},"servers":[{"url":"http://127.0.0.1:8080"}],"paths":{"/api/v1/admin/tokens":{"get":{"tags":["admin"],"operationId":"list_tokens","responses":{"200":{"description":"Device token summaries","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListTokensResponse"}}}},"500":{"description":"Token list failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenErrorResponse"}}}}}},"post":{"tags":["admin"],"operationId":"create_token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTokenRequest"}}},"required":true},"responses":{"201":{"description":"Device token minted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTokenResponse"}}}},"400":{"description":"Invalid token request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenErrorResponse"}}}},"500":{"description":"Token issuance failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenErrorResponse"}}}}}}},"/api/v1/admin/tokens/{token_id}":{"delete":{"tags":["admin"],"operationId":"revoke_token","parameters":[{"name":"token_id","in":"path","description":"Persisted device token id","required":true,"schema":{"$ref":"#/components/schemas/Id"}}],"responses":{"204":{"description":"Device token revoked"},"404":{"description":"Device token not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenErrorResponse"}}}},"500":{"description":"Token revocation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenErrorResponse"}}}}}}},"/api/v1/library/items":{"get":{"tags":["library"],"operationId":"list_catalog_items","parameters":[{"name":"type","in":"query","description":"Media item kind to include.","required":false,"schema":{"type":"string"}},{"name":"title_contains","in":"query","description":"Case-insensitive title substring.","required":false,"schema":{"type":"string"}},{"name":"has_source_file","in":"query","description":"Filter by source-file presence.","required":false,"schema":{"type":"boolean"}},{"name":"limit","in":"query","description":"Maximum number of items to return.","required":false,"schema":{"type":"integer","format":"int32","minimum":0}},{"name":"offset","in":"query","description":"Number of matching items to skip.","required":false,"schema":{"type":"integer","format":"int64","minimum":0}}],"responses":{"200":{"description":"Catalog media items","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatalogListPage"}}}},"400":{"description":"Invalid catalog list query","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/library/items/{id}/images/{kind}":{"get":{"tags":["library"],"operationId":"get_catalog_item_image","parameters":[{"name":"id","in":"path","description":"Media item id","required":true,"schema":{"$ref":"#/components/schemas/Id"}},{"name":"kind","in":"path","description":"Artwork kind: poster, backdrop, or logo","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Cached artwork image"},"401":{"description":"Missing or invalid bearer token"},"404":{"description":"Cached artwork image not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"CanonicalIdentityId":{"type":"string"},"CatalogArtwork":{"type":"object","description":"Catalog artwork URLs for one media item.","properties":{"backdrop":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/CatalogArtworkImage","description":"Backdrop artwork, when cached."}]},"logo":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/CatalogArtworkImage","description":"Logo artwork, when cached."}]},"poster":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/CatalogArtworkImage","description":"Poster artwork, when cached."}]}}},"CatalogArtworkImage":{"type":"object","description":"Catalog artwork image URLs.","properties":{"internal_url":{"type":["string","null"],"description":"Kino-owned URL clients should prefer."},"source_url":{"type":["string","null"],"description":"Provider URL used as the original image source."}}},"CatalogListPage":{"type":"object","description":"One page of catalog media items.","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/CatalogMediaItem"},"description":"Media items in this page."},"next_offset":{"type":["integer","null"],"format":"int64","description":"Offset to request the next page, when more matches exist.","minimum":0}}},"CatalogMediaItem":{"type":"object","description":"Catalog projection for one media item.","required":["id","media_kind","artwork","source_files","created_at","updated_at"],"properties":{"artwork":{"$ref":"#/components/schemas/CatalogArtwork","description":"Cached artwork URLs for this item."},"canonical_identity_id":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/CanonicalIdentityId","description":"Canonical identity for provider-backed media."}]},"created_at":{"$ref":"#/components/schemas/Timestamp","description":"Creation timestamp."},"episode_number":{"type":["integer","null"],"format":"int32","description":"TV episode number for episode rows.","minimum":0},"id":{"$ref":"#/components/schemas/Id","description":"Media item id."},"media_kind":{"$ref":"#/components/schemas/MediaItemKind","description":"Media item kind."},"season_number":{"type":["integer","null"],"format":"int32","description":"TV season number for episode rows.","minimum":0},"source_files":{"type":"array","items":{"$ref":"#/components/schemas/LibrarySourceFile"},"description":"Source files attached to this media item."},"title":{"type":["string","null"],"description":"Cached display title, when metadata has been enriched."},"updated_at":{"$ref":"#/components/schemas/Timestamp","description":"Last update timestamp."}}},"CreateTokenRequest":{"type":"object","required":["label"],"properties":{"label":{"type":"string","description":"Operator-facing label for the device token."}},"additionalProperties":false},"CreateTokenResponse":{"type":"object","required":["token","token_id","label","created_at"],"properties":{"created_at":{"$ref":"#/components/schemas/Timestamp","description":"Token creation timestamp."},"label":{"type":"string","description":"Operator-facing label for the device token."},"token":{"type":"string","description":"Plaintext bearer token. This value is returned only once."},"token_id":{"$ref":"#/components/schemas/Id","description":"Persisted device token id."}}},"ErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"Id":{"type":"string","format":"uuid"},"LibrarySourceFile":{"type":"object","description":"Persisted source-file row used for library reconciliation.","required":["id","media_item_id","path","transcode_outputs"],"properties":{"id":{"$ref":"#/components/schemas/Id","description":"Source-file id."},"media_item_id":{"$ref":"#/components/schemas/Id","description":"Media item that owns this source file."},"path":{"type":"string","description":"Canonical source path stored in the catalog."},"transcode_outputs":{"type":"array","items":{"$ref":"#/components/schemas/TranscodeOutput"},"description":"Transcode outputs derived from this source file."}}},"ListTokensResponse":{"type":"object","required":["tokens"],"properties":{"tokens":{"type":"array","items":{"$ref":"#/components/schemas/TokenSummary"},"description":"Device tokens owned by the seeded user."}}},"MediaItemKind":{"type":"string","description":"User-facing media item kind stored in the catalog.","enum":["movie","tv_episode","personal"]},"Timestamp":{"type":"string","format":"date-time"},"TokenErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"TokenSummary":{"type":"object","required":["token_id","label","created_at"],"properties":{"created_at":{"$ref":"#/components/schemas/Timestamp","description":"Token creation timestamp."},"label":{"type":"string","description":"Operator-facing label for the device token."},"last_seen_at":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Timestamp","description":"Last successful authentication time."}]},"revoked_at":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Timestamp","description":"Revocation time, if the token has been revoked."}]},"token_id":{"$ref":"#/components/schemas/Id","description":"Persisted device token id."}}},"TranscodeOutput":{"type":"object","description":"Derived stream output produced from a source file.","required":["id","source_file_id","path","created_at","updated_at"],"properties":{"created_at":{"$ref":"#/components/schemas/Timestamp","description":"Creation timestamp."},"id":{"$ref":"#/components/schemas/Id","description":"Transcode output id."},"path":{"type":"string","description":"Output path."},"source_file_id":{"$ref":"#/components/schemas/Id","description":"Source file this output was derived from."},"updated_at":{"$ref":"#/components/schemas/Timestamp","description":"Last update timestamp."}}}}},"tags":[{"name":"requests","description":"Request lifecycle operations"},{"name":"library","description":"Library catalog operations"},{"name":"admin","description":"Administrative operations"}]}